% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mint.R
\name{conv.seq}
\alias{conv.seq}
\title{Sequence convolution
conv.seq calculates the convolution of two sequences}
\usage{
conv.seq(x, y, real = TRUE)
}
\arguments{
\item{x,y}{numeric vectors}

\item{real}{logical value, indicating whether the result should be real valued}
}
\value{
a numeric vector of length equal to \code{length(x)+length(y)}, giving the convolution of \code{x} and \code{y}
}
\description{
Sequence convolution
conv.seq calculates the convolution of two sequences
}
\examples{
conv.seq(1:5,1:4)
convolve(1:5,4:1,type="open")
}
\seealso{
\code{\link{convolve}}
}

