% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sepp.resid.R
\name{sepp.resid}
\alias{sepp.resid}
\title{Calculate the self exciting point process residuals}
\usage{
sepp.resid(jtms, Tau, Inu, Ig)
}
\arguments{
\item{jtms}{A numerical vector containing the event times of the SEPP in ascending order}

\item{Tau}{A numerical scalar giving the censoring time, which should be greater than or equal to the event time}

\item{Inu}{A function that gives the integral of the baseline event rate function \eqn{\nu(t)} from 0 to the argument of the function}

\item{Ig}{A function that gives the integral of the excitation function \eqn{g(t)} from 0 to the argument of the function}
}
\value{
A numerical vector containing the SEPP residuals \eqn{\Lambda(t_i)} where \eqn{\Lambda} is the cumulative intensity process.
}
\description{
Calculate the self exciting point process residuals
}

